중간점검

+++ src/main/java/iconloop/myid/partner/adminpage/domain/repository/OrganizationRepository.java
@@ -3,5 +3,5 @@ package iconloop.myid.partner.adminpage.domain.repository;
 import iconloop.myid.partner.adminpage.domain.entity.Organization;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface OrganizationRepository extends JpaRepository<Organization, Long> {
+public interface OrganizationRepository extends JpaRepository<Organization, Integer> {
 }
